rename pages/posts to pages/feed
